pyproject.toml
[project]
name = "gradio-webview"
version = "0.1.0"
dependencies = [
    "pywebview",
    "gradio",
]
